[main] TODO check_bst_values can only fail, no pass condition

+++ BSTInputTesting/CM4/Core/Src/main.c
@@ -600,8 +600,9 @@ void VIRT_UART0_RxCpltCallback(VIRT_UART_HandleTypeDef *huart)
 
 
 
+//TODO: this only can return 
 // Checking the BST values
-int check_bst_values(float stroke, float duty_cycle1, float duty_cycle2)
+uint8_t check_bst_values(float stroke, float duty_cycle1, float duty_cycle2)
 {
 	// number of test points = 5 (25 / 5)
     int num_points = sizeof(bst_test_points) / sizeof(BSTTestPoint);
